Desacoplando, criando o BOSS e o projetil (VERSAO EM DESENVOLVIMENTO)
/* ESSA VERSAO NAO ESTA COMPLETAMENTE FUNCIONAL */

Além de criar, desacoplei algumas funções do inimigo.

Add um método virtual puro na entidade = ATACAR.

Isso pois toda entidade pode ser capaz de causar dano em outra entidade.

+++ game/Obstaculo.cpp
@@ -13,7 +13,7 @@ void Obstaculo::anulaGravidade()
 {
 	if (gravidade) {
 
-		float vetor_contrario = speed.y * -1.0;
+		double vetor_contrario = speed.y * -1.0;
 		corpo.move(Vector2f(0.0f, vetor_contrario));
 	}
